backend/requirements.txt
groq
aiohttp
aiofiles
feedparser
beautifulsoup4
html2text
python-dateutil
chromadb
sentence-transformers
yfinance
fuzzywuzzy
scipy
wikipedia-api
gTTS
fastapi
uvicorn
apscheduler
pydantic
python-dotenv
requests
pandas
pandas-ta
numpy
psycopg2-binary
sqlalchemy
pgvector
einops
